+++ agent/src/main/java/com/teamscale/jacoco/agent/commit_resolution/git_properties/GitMultiProjectPropertiesLocator.java
@@ -1,9 +1,10 @@
 package com.teamscale.jacoco.agent.commit_resolution.git_properties;
 
-import com.teamscale.jacoco.agent.options.ProjectRevision;
+import com.teamscale.jacoco.agent.options.ProjectAndCommit;
 import com.teamscale.jacoco.agent.upload.teamscale.DelayedTeamscaleMultiProjectUploader;
 import com.teamscale.jacoco.agent.util.DaemonThreadFactory;
 import com.teamscale.jacoco.agent.util.LoggingUtils;
+import org.jetbrains.annotations.VisibleForTesting;
 import org.slf4j.Logger;
 
 import java.io.File;
@@ -35,48 +36,60 @@ public class GitMultiProjectPropertiesLocator implements IGitPropertiesLocator {
 	}
 
 	public GitMultiProjectPropertiesLocator(DelayedTeamscaleMultiProjectUploader uploader, Executor executor,
-											boolean recursiveSearch) {
+			boolean recursiveSearch) {
 		this.uploader = uploader;
 		this.executor = executor;
 		this.recursiveSearch = recursiveSearch;
 	}
 
 	/**
-	 * Asynchronously searches the given jar file for a git.properties file.
+	 * Asynchronously searches the given jar file for git.properties files and adds a corresponding uploader to the
+	 * multi-project uploader.
 	 */
 	@Override
 	public void searchFileForGitPropertiesAsync(File file, boolean isJarFile) {
 		executor.execute(() -> searchFile(file, isJarFile));
 	}
 
-	private void searchFile(File file, boolean isJarFile) {
+	/**
+	 * Synchronously searches the given jar file for git.properties files and adds a corresponding uploader to the
+	 * multi-project uploader.
+	 */
+	@VisibleForTesting
+	void searchFile(File file, boolean isJarFile) {
 		logger.debug("Searching file {} for multiple git.properties", file.toString());
 		try {
-			List<ProjectRevision> projectRevisions = GitPropertiesLocatorUtils.getProjectRevisionsFromGitProperties(
+			List<ProjectAndCommit> projectAndCommits = GitPropertiesLocatorUtils.getProjectRevisionsFromGitProperties(
 					file,
 					isJarFile,
 					recursiveSearch);
-			if (projectRevisions.isEmpty()) {
+			if (projectAndCommits.isEmpty()) {
 				logger.debug("No git.properties file found in {}", file);
 				return;
 			}
-			// this code only runs when 'teamscale-project' is not given via the agent properties,
-			// i.e., a multi-project upload is being attempted.
-			// Therefore, we expect to find both the project (teamscale.project) and the revision
-			// (git.commit.id) in the git.properties file.
 
-			for (ProjectRevision projectRevision : projectRevisions) {
-				if (projectRevision.getProject() == null || projectRevision.getRevision() == null) {
-					logger.error(
+			for (ProjectAndCommit projectAndCommit : projectAndCommits) {
+				// this code only runs when 'teamscale-project' is not given via the agent properties,
+				// i.e., a multi-project upload is being attempted.
+				// Therefore, we expect to find both the project (teamscale.project) and the revision
+				// (git.commit.id) in the git.properties file.
+				if (projectAndCommit.getProject() == null || projectAndCommit.getCommitInfo() == null) {
+					logger.debug(
 							"Found inconsistent git.properties file: the git.properties file in {} either does not specify the" +
-									" Teamscale project (teamscale.project) property, or does not specify the commit SHA (git.commit.id)." +
-									" Please note that both of these properties are required in order to allow multi-project upload to Teamscale.",
-							file);
-					return;
+									" Teamscale project ({}) property, or does not specify the commit " +
+									"({}, {} + {}, or {} + {})." +
+									" Will skip this git.properties file and try to continue with the other ones that were found during discovery.",
+							file, GitPropertiesLocatorUtils.GIT_PROPERTIES_TEAMSCALE_PROJECT,
+							GitPropertiesLocatorUtils.GIT_PROPERTIES_GIT_COMMIT_ID,
+							GitPropertiesLocatorUtils.GIT_PROPERTIES_GIT_BRANCH,
+							GitPropertiesLocatorUtils.GIT_PROPERTIES_GIT_COMMIT_TIME,
+							GitPropertiesLocatorUtils.GIT_PROPERTIES_TEAMSCALE_COMMIT_BRANCH,
+							GitPropertiesLocatorUtils.GIT_PROPERTIES_TEAMSCALE_COMMIT_TIME);
+					continue;
 				}
-				uploader.setTeamscaleProjectForRevision(projectRevision);
+				uploader.setTeamscaleProjectForRevision(projectAndCommit);
 				logger.debug("Found git.properties file in {} and found Teamscale project {} and revision {}", file,
-						projectRevision.getProject(), projectRevision.getRevision());
+						projectAndCommit.getProject(), projectAndCommit.getCommitInfo());
 			}
 		} catch (IOException | InvalidGitPropertiesException e) {
 			logger.error("Error during asynchronous search for git.properties in {}", file, e);