+++ agent/src/main/java/com/teamscale/jacoco/agent/upload/teamscale/DelayedTeamscaleMultiProjectUploader.java
@@ -1,9 +1,9 @@
 package com.teamscale.jacoco.agent.upload.teamscale;
 
-import com.teamscale.jacoco.agent.options.AgentOptions;
-import com.teamscale.jacoco.agent.options.ProjectRevision;
-import com.teamscale.jacoco.agent.upload.IUploader;
+import com.teamscale.jacoco.agent.commit_resolution.git_properties.CommitInfo;
+import com.teamscale.jacoco.agent.options.ProjectAndCommit;
 import com.teamscale.jacoco.agent.upload.DelayedMultiUploaderBase;
+import com.teamscale.jacoco.agent.upload.IUploader;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -13,16 +13,17 @@ import java.util.function.BiFunction;
 /** Wrapper for {@link TeamscaleUploader} that allows to upload the same coverage file to multiple Teamscale projects. */
 public class DelayedTeamscaleMultiProjectUploader extends DelayedMultiUploaderBase implements IUploader {
 
-	private final BiFunction<String, String, IUploader> uploaderFactory;
+	private final BiFunction<String, CommitInfo, IUploader> uploaderFactory;
 	private final List<IUploader> teamscaleUploaders = new ArrayList<>();
 
-	public DelayedTeamscaleMultiProjectUploader(BiFunction<String, String, IUploader> uploaderFactory) {
+	public DelayedTeamscaleMultiProjectUploader(BiFunction<String, CommitInfo, IUploader> uploaderFactory) {
 		this.uploaderFactory = uploaderFactory;
 	}
 
 	/** Sets the project and revision detected for the Teamscale project. */
-	public void setTeamscaleProjectForRevision(ProjectRevision projectRevision) {
-		IUploader uploader = uploaderFactory.apply(projectRevision.getProject(), projectRevision.getRevision());
+	public void setTeamscaleProjectForRevision(ProjectAndCommit projectAndCommit) {
+		IUploader uploader = uploaderFactory.apply(projectAndCommit.getProject(),
+				projectAndCommit.getCommitInfo());
 		teamscaleUploaders.add(uploader);
 	}